Add controller class Add shape interface

+++ api/src/main/java/it/unicam/cs/pa/api/Point.java
@@ -6,5 +6,5 @@ package it.unicam.cs.pa.api;
  * @param x the x coordinate
  * @param y the y coordinate
  */
-public record Point(int x, int y) {
+public record Point(double x, double y) {
 }
